fix: make reusable useFileURLCopy hook
